Restructure IFrontController for better performance and usability
