--- src/xunit.v3.runner.utility.tests/Frameworks/v2/Xunit2Tests.cs
@@ -1,324 +1,0 @@
-﻿#if NETFRAMEWORK
-
-using System.Linq;
-using Xunit;
-using Xunit.Runner.Common;
-using Xunit.Runner.v2;
-using Xunit.v3;
-
-public class Xunit2Tests
-{
-	public class EnumerateTests
-	{
-		[Fact]
-		public async void NoTestMethods()
-		{
-			using var assm = await CSharpAcceptanceTestV2Assembly.Create(code: "");
-			var controller = new TestableXunit2(assm.FileName, null, true);
-			using var sink = SpyMessageSink<_DiscoveryComplete>.Create();
-			var settings = new FrontControllerFindSettings(_TestFrameworkOptions.ForDiscovery());
-
-			controller.Find(sink, settings);
-
-			sink.Finished.WaitOne();
-
-			Assert.IsType<_DiscoveryStarting>(sink.Messages.First());
-			Assert.False(sink.Messages.Any(msg => msg is _TestCaseDiscovered));
-		}
-
-		[Fact]
-		public async void SingleTestMethod()
-		{
-			var code = @"
-using Xunit;
-
-public class Foo
-{
-	[Fact]
-	public void Bar() { }
-}";
-
-			using var assm = await CSharpAcceptanceTestV2Assembly.Create(code);
-			var controller = new TestableXunit2(assm.FileName, null, true);
-			using var sink = SpyMessageSink<_DiscoveryComplete>.Create();
-			var settings = new FrontControllerFindSettings(_TestFrameworkOptions.ForDiscovery());
-
-			controller.Find(sink, settings);
-
-			sink.Finished.WaitOne();
-
-			var testCase = sink.Messages.OfType<_TestCaseDiscovered>().Single();
-			Assert.Equal("Foo.Bar", testCase.TestCaseDisplayName);
-		}
-	}
-
-	public class CSharp
-	{
-		[Fact]
-		public async void FactAcceptanceTest_Find()
-		{
-			var code = @"
-using System;
-using Xunit;
-
-namespace Namespace1
-{
-	public class Class1
-	{
-		[Fact]
-		[Trait(""Name!"", ""Value!"")]
-		public void Trait() { }
-
-		[Fact(Skip=""Skipping"")]
-		public void Skipped() { }
-
-		[Fact(DisplayName=""Custom Test Name"")]
-		public void CustomName() { }
-	}
-}
-
-namespace Namespace2
-{
-	public class OuterClass
-	{
-		public class Class2
-		{
-			[Fact]
-			public void TestMethod() { }
-		}
-	}
-}";
-
-			using var assembly = await CSharpAcceptanceTestV2Assembly.Create(code);
-			var controller = new TestableXunit2(assembly.FileName, null, true);
-			using var sink = SpyMessageSink<_DiscoveryComplete>.Create();
-			var settings = new FrontControllerFindSettings(_TestFrameworkOptions.ForDiscovery());
-
-			controller.Find(sink, settings);
-
-			sink.Finished.WaitOne();
-			var testCases = sink.Messages.OfType<_TestCaseDiscovered>().ToArray();
-
-			Assert.Equal(4, testCases.Length);
-
-			var innerClassTest = Assert.Single(testCases, tc => tc.TestCaseDisplayName == "Namespace2.OuterClass+Class2.TestMethod");
-			Assert.StartsWith(":F:Namespace2.OuterClass+Class2:TestMethod:1:0:", innerClassTest.Serialization);
-			Assert.Null(innerClassTest.SkipReason);
-			Assert.Equal("Class2", innerClassTest.TestClass);
-			Assert.Equal("Namespace2.OuterClass+Class2", innerClassTest.TestClassWithNamespace);
-			Assert.Equal("TestMethod", innerClassTest.TestMethod);
-			Assert.Equal("Namespace2", innerClassTest.TestNamespace);
-
-			var traitTest = Assert.Single(testCases, tc => tc.TestCaseDisplayName == "Namespace1.Class1.Trait");
-			var key = Assert.Single(traitTest.Traits.Keys);
-			Assert.Equal("Name!", key);
-			var value = Assert.Single(traitTest.Traits[key]);
-			Assert.Equal("Value!", value);
-
-			var skipped = Assert.Single(testCases, tc => tc.TestCaseDisplayName == "Namespace1.Class1.Skipped");
-			Assert.Equal("Skipping", skipped.SkipReason);
-
-			Assert.Single(testCases, tc => tc.TestCaseDisplayName == "Custom Test Name");
-		}
-
-		[Fact]
-		public async void FactAcceptanceTest_Run()
-		{
-			var code = @"
-using System;
-using Xunit;
-
-public class TestClass
-{
-	[Fact]
-	public void TestMethod() { Assert.True(false); }
-}";
-
-			using var assembly = await CSharpAcceptanceTestV2Assembly.Create(code);
-			var controller = new TestableXunit2(assembly.FileName, null, true);
-			var discoveryOptions = _TestFrameworkOptions.ForDiscovery();
-			var settings = new FrontControllerFindSettings(discoveryOptions);
-
-			using var discoverySink = SpyMessageSink<_DiscoveryComplete>.Create();
-			controller.Find(discoverySink, settings);
-			discoverySink.Finished.WaitOne();
-
-			using var executionSink = SpyMessageSink<_TestAssemblyFinished>.Create();
-
-			var serializedTestCases = discoverySink.Messages.OfType<_TestCaseDiscovered>().Select(tcdm => tcdm.Serialization!).ToArray();
-			Assert.All(serializedTestCases, serializedTestCase => Assert.NotNull(serializedTestCase));
-			controller.RunTests(serializedTestCases, executionSink, _TestFrameworkOptions.ForExecution());
-
-			executionSink.Finished.WaitOne();
-
-			Assert.Empty(executionSink.Messages.OfType<_TestPassed>());
-			Assert.Empty(executionSink.Messages.OfType<_TestSkipped>());
-			var failedTest = Assert.Single(executionSink.Messages.OfType<_TestFailed>());
-			var failedMetadata = executionSink.Messages.OfType<_TestStarting>().Single(ts => ts.TestUniqueID == failedTest.TestUniqueID);
-			Assert.Equal("TestClass.TestMethod", failedMetadata.TestDisplayName);
-		}
-
-		[Fact]
-		public async void TheoryWithInlineData()
-		{
-			var code = @"
-using System;
-using Xunit;
-
-public class TestClass
-{
-	[Theory]
-	[InlineData]
-	[InlineData(42)]
-	[InlineData(42, 21.12)]
-	public void TestMethod(int x) { }
-}";
-
-			using var assembly = await CSharpAcceptanceTestV2Assembly.Create(code);
-			var controller = new TestableXunit2(assembly.FileName, null, true);
-			using var sink = SpyMessageSink<_DiscoveryComplete>.Create();
-			var settings = new FrontControllerFindSettings(_TestFrameworkOptions.ForDiscovery());
-
-			controller.Find(sink, settings);
-
-			sink.Finished.WaitOne();
-			var testCaseNames = sink.Messages.OfType<_TestCaseDiscovered>().Select(tcdm => tcdm.TestCaseDisplayName).ToArray();
-
-			Assert.Equal(3, testCaseNames.Length);
-
-			Assert.Contains("TestClass.TestMethod(x: ???)", testCaseNames);
-			Assert.Contains("TestClass.TestMethod(x: 42)", testCaseNames);
-			Assert.Contains($"TestClass.TestMethod(x: 42, ???: {21.12})", testCaseNames);
-		}
-	}
-
-	public class FSharp
-	{
-		[Fact]
-		public async void FactAcceptanceTest()
-		{
-			var code = @"
-module FSharpTests
-
-open Xunit
-
-[<Fact>]
-[<Trait(""Name!"", ""Value!"")>]
-let Trait() =
-    Assert.True(true)
-
-[<Fact(Skip = ""Skipping"")>]
-let Skipped() =
-    Assert.True(false)
-
-[<Fact(DisplayName=""Custom Test Name"")>]
-let CustomName() =
-    Assert.True(true)
-";
-
-			using var assembly = await FSharpAcceptanceTestV2Assembly.Create(code);
-			var controller = new TestableXunit2(assembly.FileName, null, true);
-			var sink = new TestDiscoverySink();
-			var settings = new FrontControllerFindSettings(_TestFrameworkOptions.ForDiscovery());
-
-			controller.Find(sink, settings);
-
-			sink.Finished.WaitOne();
-
-			Assert.Collection(
-				sink.TestCases.OrderBy(tc => tc.TestCaseDisplayName),
-				testCase => Assert.Equal("Custom Test Name", testCase.TestCaseDisplayName),
-				testCase =>
-				{
-					Assert.Equal("FSharpTests.Skipped", testCase.TestCaseDisplayName);
-					Assert.Equal("Skipping", testCase.SkipReason);
-				},
-				testCase =>
-				{
-					Assert.Equal("FSharpTests.Trait", testCase.TestCaseDisplayName);
-					Assert.Collection(testCase.Traits,
-						kvp =>
-						{
-							Assert.Equal("Name!", kvp.Key);
-							Assert.Equal("Value!", kvp.Value.Single());
-						}
-					);
-				}
-			);
-		}
-
-		[Fact]
-		public async void TheoryWithInlineData()
-		{
-			var code = @"
-module FSharpTests
-
-open Xunit
-
-[<Theory>]
-[<InlineData>]
-[<InlineData(42)>]
-[<InlineData(42, 21.12)>]
-let TestMethod (x:int) =
-    Assert.True(true)
-";
-
-			using var assembly = await FSharpAcceptanceTestV2Assembly.Create(code);
-			var controller = new TestableXunit2(assembly.FileName, null, true);
-			var sink = new TestDiscoverySink();
-			var settings = new FrontControllerFindSettings(_TestFrameworkOptions.ForDiscovery());
-
-			controller.Find(sink, settings);
-			sink.Finished.WaitOne();
-
-			Assert.Collection(
-				sink.TestCases.OrderBy(tc => tc.TestCaseDisplayName),
-				testCase => Assert.Equal("FSharpTests.TestMethod(x: ???)", testCase.TestCaseDisplayName),
-				testCase => Assert.Equal("FSharpTests.TestMethod(x: 42)", testCase.TestCaseDisplayName),
-				testCase => Assert.Equal("FSharpTests.TestMethod(x: 42, ???: 21.12)", testCase.TestCaseDisplayName)
-			);
-		}
-
-		[Fact]
-		public async void SupportsAsyncReturningMethods()
-		{
-			var code = @"
-module FSharpTests
-
-open Xunit
-
-[<Fact>]
-let AsyncFailing() =
-    async {
-        do! Async.Sleep(10)
-        Assert.True(false)
-    }
-";
-
-			using var assembly = await FSharpAcceptanceTestV2Assembly.Create(code);
-			var controller = new TestableXunit2(assembly.FileName, null, true);
-			using var sink = SpyMessageSink<_TestAssemblyFinished>.Create();
-
-			controller.RunAll(sink, discoveryOptions: _TestFrameworkOptions.ForDiscovery(), executionOptions: _TestFrameworkOptions.ForExecution());
-			sink.Finished.WaitOne();
-
-			var failures = sink.Messages.OfType<_TestFailed>();
-			var failure = Assert.Single(failures);
-			var failureStarting = sink.Messages.OfType<_TestStarting>().Single(s => s.TestUniqueID == failure.TestUniqueID);
-			Assert.Equal("FSharpTests.AsyncFailing", failureStarting.TestDisplayName);
-		}
-	}
-
-	class TestableXunit2 : Xunit2
-	{
-		public TestableXunit2(
-			string assemblyFileName,
-			string? configFileName = null,
-			bool shadowCopy = true,
-			AppDomainSupport appDomainSupport = AppDomainSupport.Required)
-				: base(new _NullMessageSink(), appDomainSupport, _NullSourceInformationProvider.Instance, assemblyFileName, configFileName, shadowCopy)
-		{ }
-	}
-}
-
-#endif
