+++ src/xunit.v3.runner.utility/Frameworks/v1/Xunit1.cs
@@ -102,13 +102,11 @@ namespace Xunit.Runner.v1
 		protected virtual IXunit1Executor CreateExecutor() =>
 			new Xunit1Executor(diagnosticMessageSink, appDomainSupport != AppDomainSupport.Denied, assemblyFileName, configFileName, shadowCopy, shadowCopyFolder);
 
-		/// <inheritdoc/>
-		public Xunit1TestCase? Deserialize(string value)
-		{
-			Guard.ArgumentNotNull(nameof(value), value);
-
-			return SerializationHelper.Deserialize<Xunit1TestCase>(value);
-		}
+		/// <summary>
+		/// INTERNAL METHOD, FOR TESTING PURPOSES ONLY. DO NOT CALL.
+		/// </summary>
+		protected Xunit1TestCase? Deserialize(string value) =>
+			SerializationHelper.Deserialize<Xunit1TestCase>(value);
 
 		/// <inheritdoc/>
 		public async ValueTask DisposeAsync()
@@ -123,64 +121,15 @@ namespace Xunit.Runner.v1
 		}
 
 		/// <summary>
-		/// Starts the process of finding all xUnit.net v1 tests in an assembly.
+		/// INTERNAL METHOD, FOR TESTING PURPOSES ONLY. DO NOT CALL.
 		/// </summary>
-		/// <param name="includeSourceInformation">Whether to include source file information, if possible.</param>
-		/// <param name="messageSink">The message sink to report results back to.</param>
-		/// <param name="filters">The optional filters used to filter discovery messages.</param>
-		public void Find(
-			bool includeSourceInformation,
+		protected void Find(
 			_IMessageSink messageSink,
-			XunitFilters? filters = null)
-		{
-			Guard.ArgumentNotNull(nameof(messageSink), messageSink);
-
-			Predicate<_TestCaseDiscovered> filter = filters != null ? filters.Filter : msg => true;
-
-			Find(filter, includeSourceInformation, messageSink);
-		}
-
-		/// <summary>
-		/// Starts the process of finding all xUnit.net v1 tests in a class.
-		/// </summary>
-		/// <param name="typeName">The fully qualified type name to find tests in.</param>
-		/// <param name="includeSourceInformation">Whether to include source file information, if possible.</param>
-		/// <param name="messageSink">The message sink to report results back to.</param>
-		public void Find(
-			string typeName,
 			bool includeSourceInformation,
-			_IMessageSink messageSink)
-		{
-			Guard.ArgumentNotNullOrEmpty(nameof(typeName), typeName);
-			Guard.ArgumentNotNull(nameof(messageSink), messageSink);
-
-			Find(
-				msg => msg.TestClassWithNamespace == typeName,
-				includeSourceInformation,
-				messageSink
-			);
-		}
-
-		/// <inheritdoc/>
-		void IFrontController.Find(
-			_IMessageSink messageSink,
-			FrontControllerFindSettings settings)
+			Predicate<_TestCaseDiscovered>? filter)
 		{
 			Guard.ArgumentNotNull(nameof(messageSink), messageSink);
-			Guard.ArgumentNotNull(nameof(settings), settings);
 
-			Find(
-				settings.Filters.Filter,
-				settings.Options.GetIncludeSourceInformationOrDefault(),
-				messageSink
-			);
-		}
-
-		void Find(
-			Predicate<_TestCaseDiscovered> filter,
-			bool includeSourceInformation,
-			_IMessageSink messageSink)
-		{
 			var discoveryStarting = new _DiscoveryStarting
 			{
 				AssemblyName = testAssemblyName,
@@ -192,105 +141,12 @@ namespace Xunit.Runner.v1
 
 			try
 			{
-				XmlNode? assemblyXml = null;
-
-				var handler = new XmlNodeCallbackHandler(xml => { assemblyXml = xml; return true; });
-				Executor.EnumerateTests(handler);
-
-				if (assemblyXml != null)
+				Find(includeSourceInformation, testCase =>
 				{
-					var methodNodes = assemblyXml.SelectNodes("//method")?.Cast<XmlNode>();
-					if (methodNodes != null)
-					{
-						foreach (var methodXml in methodNodes)
-						{
-							var typeName = methodXml.Attributes?["type"]?.Value;
-							var methodName = methodXml.Attributes?["method"]?.Value;
-							if (typeName == null || methodName == null)
-								continue;
-
-							string? displayName = null;
-							var displayNameAttribute = methodXml.Attributes?["name"];
-							if (displayNameAttribute != null)
-								displayName = displayNameAttribute.Value;
-
-							string? skipReason = null;
-							var skipReasonAttribute = methodXml.Attributes?["skip"];
-							if (skipReasonAttribute != null)
-								skipReason = skipReasonAttribute.Value;
-
-							var traits = new Dictionary<string, List<string>>(StringComparer.OrdinalIgnoreCase);
-							var traitNodes = methodXml.SelectNodes("traits/trait")?.Cast<XmlNode>();
-							if (traitNodes != null)
-								foreach (var traitNode in traitNodes)
-								{
-									var traitName = traitNode.Attributes?["name"]?.Value;
-									var traitValue = traitNode.Attributes?["value"]?.Value;
-
-									if (traitName != null && traitValue != null)
-										traits.Add(traitName, traitValue);
-								}
-
-							var sourceInformation = default(_ISourceInformation);
-							if (includeSourceInformation)
-								sourceInformation = sourceInformationProvider.GetSourceInformation(typeName, methodName);
-
-							string? @namespace = null;
-							string? @class = null;
-
-							var namespaceIdx = typeName.LastIndexOf('.');
-							if (namespaceIdx < 0)
-								@class = typeName;
-							else
-							{
-								@namespace = typeName.Substring(0, namespaceIdx);
-								@class = typeName.Substring(namespaceIdx + 1);
-
-								var innerClassIdx = @class.LastIndexOf('+');
-								if (innerClassIdx >= 0)
-									@class = @class.Substring(innerClassIdx + 1);
-							}
-
-							var testCase = new Xunit1TestCase
-							{
-								AssemblyUniqueID = testAssemblyUniqueID,
-								SkipReason = skipReason,
-								SourceFilePath = sourceInformation?.FileName,
-								SourceLineNumber = sourceInformation?.LineNumber,
-								TestCaseDisplayName = displayName ?? $"{typeName}.{methodName}",
-								TestCaseUniqueID = $":v1:case:{typeName}.{methodName}:{assemblyFileName}:{configFileName ?? "(null)"}",
-								TestClass = typeName,
-								TestClassUniqueID = $":v1:class:{typeName}:{assemblyFileName}:{configFileName ?? "(null)"}",
-								TestCollectionUniqueID = $":v1:collection:{assemblyFileName}:{configFileName ?? "(null)"}",
-								TestMethod = methodName,
-								TestMethodUniqueID = $":v1:method:{typeName}.{methodName}:{assemblyFileName}:{configFileName ?? "(null)"}",
-								Traits = traits
-							};
-
-							var message = new _TestCaseDiscovered
-							{
-								AssemblyUniqueID = testCase.AssemblyUniqueID,
-								Serialization = SerializationHelper.Serialize(testCase),
-								SkipReason = testCase.SkipReason,
-								SourceFilePath = testCase.SourceFilePath,
-								SourceLineNumber = testCase.SourceLineNumber,
-								TestCaseDisplayName = testCase.TestCaseDisplayName,
-								TestCaseUniqueID = testCase.TestCaseUniqueID,
-								TestClass = @class,
-								TestClassUniqueID = testCase.TestClassUniqueID,
-								TestClassWithNamespace = testCase.TestClass,
-								TestCollectionUniqueID = testCase.TestCollectionUniqueID,
-								TestMethod = testCase.TestMethod,
-								TestMethodUniqueID = testCase.TestMethodUniqueID,
-								TestNamespace = @namespace,
-								Traits = testCase.Traits
-							};
-
-							if (filter(message))
-								messageSink.OnMessage(message);
-						}
-					}
-				}
+					var msg = testCase.ToTestCaseDiscovered(includeSerialization: true);
+					if (filter == null || filter(msg))
+						messageSink.OnMessage(msg);
+				});
 			}
 			finally
 			{
@@ -299,72 +155,162 @@ namespace Xunit.Runner.v1
 			}
 		}
 
-		/// <summary>
-		/// Starts the process of running all the xUnit.net v1 tests in the assembly.
-		/// </summary>
-		/// <param name="messageSink">The message sink to report results back to.</param>
-		public void Run(_IMessageSink messageSink)
-		{
-			var discoverySink = new TestDiscoverySink();
-			disposalTracker.Add(discoverySink);
-
-			Find(includeSourceInformation: false, discoverySink);
-			discoverySink.Finished.WaitOne();
-
-			Run(discoverySink.TestCases.Select(tc => tc.Serialization), messageSink);
-		}
-
-		void IFrontController.RunAll(
+		/// <inheritdoc/>
+		public void Find(
 			_IMessageSink messageSink,
-			_ITestFrameworkDiscoveryOptions discoveryOptions,
-			_ITestFrameworkExecutionOptions executionOptions)
+			FrontControllerFindSettings settings)
 		{
-			Run(messageSink);
+			Guard.ArgumentNotNull(nameof(messageSink), messageSink);
+			Guard.ArgumentNotNull(nameof(settings), settings);
+
+			Find(
+				messageSink,
+				settings.Options.GetIncludeSourceInformationOrDefault(),
+				settings.Filters.Empty ? null : settings.Filters.Filter
+			);
+		}
+
+		void Find(
+			bool includeSourceInformation,
+			Action<Xunit1TestCase> callback)
+		{
+			XmlNode? assemblyXml = null;
+
+			var handler = new XmlNodeCallbackHandler(xml => { assemblyXml = xml; return true; });
+			Executor.EnumerateTests(handler);
+
+			if (assemblyXml != null)
+			{
+				var methodNodes = assemblyXml.SelectNodes("//method")?.Cast<XmlNode>();
+				if (methodNodes != null)
+				{
+					foreach (var methodXml in methodNodes)
+					{
+						var typeName = methodXml.Attributes?["type"]?.Value;
+						var methodName = methodXml.Attributes?["method"]?.Value;
+						if (typeName == null || methodName == null)
+							continue;
+
+						string? displayName = null;
+						var displayNameAttribute = methodXml.Attributes?["name"];
+						if (displayNameAttribute != null)
+							displayName = displayNameAttribute.Value;
+
+						string? skipReason = null;
+						var skipReasonAttribute = methodXml.Attributes?["skip"];
+						if (skipReasonAttribute != null)
+							skipReason = skipReasonAttribute.Value;
+
+						var traits = new Dictionary<string, List<string>>(StringComparer.OrdinalIgnoreCase);
+						var traitNodes = methodXml.SelectNodes("traits/trait")?.Cast<XmlNode>();
+						if (traitNodes != null)
+							foreach (var traitNode in traitNodes)
+							{
+								var traitName = traitNode.Attributes?["name"]?.Value;
+								var traitValue = traitNode.Attributes?["value"]?.Value;
+
+								if (traitName != null && traitValue != null)
+									traits.Add(traitName, traitValue);
+							}
+
+						var sourceInformation = default(_ISourceInformation);
+						if (includeSourceInformation)
+							sourceInformation = sourceInformationProvider.GetSourceInformation(typeName, methodName);
+
+						var testCase = new Xunit1TestCase
+						{
+							AssemblyUniqueID = testAssemblyUniqueID,
+							SkipReason = skipReason,
+							SourceFilePath = sourceInformation?.FileName,
+							SourceLineNumber = sourceInformation?.LineNumber,
+							TestCaseDisplayName = displayName ?? $"{typeName}.{methodName}",
+							TestCaseUniqueID = $":v1:case:{typeName}.{methodName}:{assemblyFileName}:{configFileName ?? "(null)"}",
+							TestClass = typeName,
+							TestClassUniqueID = $":v1:class:{typeName}:{assemblyFileName}:{configFileName ?? "(null)"}",
+							TestCollectionUniqueID = $":v1:collection:{assemblyFileName}:{configFileName ?? "(null)"}",
+							TestMethod = methodName,
+							TestMethodUniqueID = $":v1:method:{typeName}.{methodName}:{assemblyFileName}:{configFileName ?? "(null)"}",
+							Traits = traits
+						};
+
+						callback(testCase);
+					}
+				}
+			}
 		}
 
 		/// <summary>
-		/// Starts the process of running all the xUnit.net v1 tests.
+		/// INTERNAL METHOD, FOR TESTING PURPOSES ONLY. DO NOT CALL.
 		/// </summary>
-		/// <param name="serializedTestCases">The serialized test cases to run</param>
-		/// <param name="messageSink">The message sink to report results back to.</param>
-		public void Run(
-			IEnumerable<string> serializedTestCases,
-			_IMessageSink messageSink)
+		protected void FindAndRun(
+			_IMessageSink messageSink,
+			bool includeSourceInformation,
+			Predicate<_TestCaseDiscovered>? filter)
 		{
-			var testCases = serializedTestCases.Select(x => Deserialize(x)).WhereNotNull().ToList();
+			Guard.ArgumentNotNull(nameof(messageSink), messageSink);
+
+			var testCases = new List<Xunit1TestCase>();
+
+			Find(includeSourceInformation, testCase =>
+			{
+				var include = true;
+
+				if (filter != null)
+				{
+					var msg = testCase.ToTestCaseDiscovered(includeSerialization: false);
+					include = filter(msg);
+				}
+
+				if (include)
+					testCases.Add(testCase);
+			});
+
 			Run(testCases, messageSink);
 		}
 
+		/// <inheritdoc/>
+		public void FindAndRun(
+			_IMessageSink messageSink,
+			FrontControllerFindAndRunSettings settings)
+		{
+			Guard.ArgumentNotNull(nameof(messageSink), messageSink);
+			Guard.ArgumentNotNull(nameof(settings), settings);
+
+			// Pass null for empty filter list, since it bypasses _TestCaseDiscovered creation
+			FindAndRun(
+				messageSink,
+				settings.DiscoveryOptions.GetIncludeSourceInformationOrDefault(),
+				settings.Filters.Empty ? null : settings.Filters.Filter
+			);
+		}
+
 		/// <summary>
-		/// Starts the process of running all the xUnit.net v1 tests.
+		/// INTERNAL METHOD, FOR TESTING PURPOSES ONLY. DO NOT CALL.
 		/// </summary>
-		/// <param name="testCases">The test cases to run; if null, all tests in the assembly are run.</param>
-		/// <param name="messageSink">The message sink to report results back to.</param>
-		public void Run(
+		protected void Run(
 			IEnumerable<Xunit1TestCase> testCases,
 			_IMessageSink messageSink)
 		{
 			var results = new Xunit1RunSummary();
 			var environment = $"{IntPtr.Size * 8}-bit .NET {Environment.Version}";
-
 			var testCasesList = testCases.ToList();
 
-			if (testCasesList.Count != 0)
+			var testAssemblyStartingMessage = new _TestAssemblyStarting
 			{
-				var testAssemblyStartingMessage = new _TestAssemblyStarting
-				{
-					AssemblyName = testAssemblyName,
-					AssemblyPath = assemblyFileName,
-					AssemblyUniqueID = testCasesList[0].AssemblyUniqueID,
-					ConfigFilePath = configFileName,
-					StartTime = DateTimeOffset.Now,
-					TestEnvironment = environment,
-					TestFrameworkDisplayName = TestFrameworkDisplayName,
-				};
+				AssemblyName = testAssemblyName,
+				AssemblyPath = assemblyFileName,
+				AssemblyUniqueID = testAssemblyUniqueID,
+				ConfigFilePath = configFileName,
+				StartTime = DateTimeOffset.Now,
+				TestEnvironment = environment,
+				TestFrameworkDisplayName = TestFrameworkDisplayName,
+			};
 
+			if (messageSink.OnMessage(testAssemblyStartingMessage))
+			{
 				try
 				{
-					if (messageSink.OnMessage(testAssemblyStartingMessage))
+					if (testCasesList.Count != 0)
 						results = RunTestCollection(testCasesList, messageSink);
 				}
 				catch (Exception ex)
@@ -395,11 +341,16 @@ namespace Xunit.Runner.v1
 			}
 		}
 
-		void IFrontController.RunTests(
-			IEnumerable<string> serializedTestCases,
-			_IMessageSink executionMessageSink,
-			_ITestFrameworkExecutionOptions executionOptions) =>
-				Run(serializedTestCases, executionMessageSink);
+		/// <inheritdoc/>
+		public void Run(
+			_IMessageSink messageSink,
+			FrontControllerRunSettings settings)
+		{
+			Guard.ArgumentNotNull(nameof(messageSink), messageSink);
+			Guard.ArgumentNotNull(nameof(settings), settings);
+
+			Run(settings.SerializedTestCases.Select(tc => Deserialize(tc)).WhereNotNull(), messageSink);
+		}
 
 		Xunit1RunSummary RunTestCollection(
 			IList<Xunit1TestCase> testCases,
@@ -495,6 +446,12 @@ namespace Xunit.Runner.v1
 
 			return results;
 		}
+
+		/// <summary>
+		/// INTERNAL METHOD, FOR TESTING PURPOSES ONLY. DO NOT CALL.
+		/// </summary>
+		protected string Serialize(Xunit1TestCase testCase) =>
+			SerializationHelper.Serialize(testCase);
 	}
 }
